On the way to working out fading properly
There’s a logic fail on the code to fade colours. This isn’t behaving
properly, and instead converges on a value that’s removed from the
desired colour (because  the measured difference decreases).

The Arduino requires external pullups to prevent the interrupts from
firing, so disabled these whilst testing the rest of the code (testing
will be required).
Some more debug stuff added — can be removed once it’s all working.