Fading working
The logic fail in the fading code was that the numbers were all
integers, so rounding errors compounded one another, and we ended up
with something that was off by quite a bit (since in the test cases,
the delta was 3.99, all those 0.99’s added up to a large incorrect
value).

Moved some of the RTC handing stuff (including code that gets the time
from the user) into a separate header file to try to manage the length
of the main code, which was becoming unmanageably long.

Added some switches for turning debugging on and off.

Probably needs some rationalising of the locations of the stuff in the
header files, but it’s passable for now