Working standalone  alarm version
The up and down buttons aren’t working yet, but I suspect that my
hardware is the issue there, rather than something that I’m doing with
the software since a simple thing to read button states is also not
working as expected.
Removed the blocking call to !Serial at the start since we’re using a
Leonardo here, and without the USB connected, it’ll do nothing with
this line present.